refactor(studio): migrate sessions+signals routers onto registry (phase 2 area 1)
Move GET /sessions/, GET /sessions/{id}, GET /sessions/{id}/stream, and
GET /sessions/{id}/signals handlers from routers/sessions.py and
routers/signals.py into services/sessions.py decorated with @studio_route.

Both SSE stream handlers (stream_session_route, stream_signals) preserve
the generate() closures and heartbeat/done frames verbatim. The signals
handler is co-located in services/sessions.py because it shares the
session existence check and stream-done logic from that module.

Rewire test_signals_sse.py to import stream_signals from services.sessions
instead of the deleted routers.signals module.

Oracle: routes.json identical, openapi.json identical (79 routes).
